rename dashboard to sidenav
